Add gravitational wave (GW) mode support and fix numerical stability issues
Major changes:
- Converted pipeline from particle physics to gravitational wave anomaly detection
- Added GW data generation (gen_data.py, gw_processing.py) with synthetic signals
- Updated config.py to use time windows (ms) instead of energy (TeV)
- Modified preprocessing.py to generate GW signals/backgrounds dynamically
- Fixed MPS (Apple Silicon) float64 compatibility in model_S.py and train_model_S.py
- Fixed numerical stability in fitting.py (added epsilon to prevent log(0) errors)
- Reduced training epochs for faster testing (bkgtemplate.py, rnodetemplate.py)
- Simplified pipeline defaults in law.py (num templates 1 instead of 20)

This enables running RANODE anomaly detection on gravitational wave time-series
data on Mac M-series GPUs. Successfully tested on M2 Pro with MPS acceleration.

+++ ranode/src/models/train_model_S.py
@@ -80,11 +80,12 @@ def train_model_S(
     traintensor_B = torch.from_numpy(data_train_SR_B.astype("float32")).to(device)
     valtensor_B = torch.from_numpy(data_val_SR_B.astype("float32")).to(device)
     # p(m) for bkg model p(x|m)
+    # Convert to float32 for MPS compatibility (Apple Silicon doesn't support float64)
     train_mass_prob_B = torch.from_numpy(
-        density_back.pdf(traintensor_B[:, 0].cpu().detach().numpy())
+        density_back.pdf(traintensor_B[:, 0].cpu().detach().numpy()).astype("float32")
     ).to(device)
     val_mass_prob_B = torch.from_numpy(
-        density_back.pdf(valtensor_B[:, 0].cpu().detach().numpy())
+        density_back.pdf(valtensor_B[:, 0].cpu().detach().numpy()).astype("float32")
     ).to(device)
 
     print("data loaded")
